Get a pseudobulk of bins with at least 2 replicates

This function will try to create a pseudobulked count matrix for the bins. When a replicate has too few cells, it is discounted. If only one exists, then we sample from it twice to create the pseudobulks.

Usage

get_bins_as_bulk(
  pseudotime_sce,
  min_cells_for_bulk = 50,
  replicate_slot = "replicate"
)

Arguments

pseudotime_sce

The Single Cell Experiment object to get the bins from

min_cells_for_bulk

The minimum cells to look for per replicate and bin.

replicate_slot

The slot in the Single Cell Experiment that contains replicate information

Value

A dataframe pseudobulk counts matrix.

Examples

library(SingleCellExperiment, quietly=TRUE)
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
library(atgnat)
counts <- matrix(rpois(1000, lambda = 10), ncol=100, nrow=10)
sce <- SingleCellExperiment::SingleCellExperiment(
  assays = list(normcounts = counts, counts = counts/2)
)
sce$pseudotime = seq_len(100)
colnames(sce) = seq_len(100)
rownames(sce) = as.character(seq_len(10))
sce = assign_pseudotime_bins(sce, n_bins=5,
  pseudotime_slot="pseudotime", split_by="cells")
sce$replicate=rep(c(1,2), 50)
result = get_bins_as_bulk(sce, min_cells_for_bulk=1, replicate_slot="replicate")
result
#>    bin_1_rep_1 bin_1_rep_2 bin_2_rep_1 bin_2_rep_2 bin_3_rep_1 bin_3_rep_2
#> 1         45.5        42.0        45.0        50.5        59.0        46.5
#> 10        54.0        51.5        56.5        42.5        52.0        50.0
#> 2         52.5        49.0        46.0        52.5        48.5        56.5
#> 3         53.5        50.5        43.5        51.5        43.5        47.0
#> 4         44.0        44.0        50.5        40.5        50.0        45.5
#> 5         47.0        45.0        50.5        44.0        50.5        49.0
#> 6         54.0        45.0        42.5        55.5        55.5        55.0
#> 7         50.0        50.0        53.5        42.5        44.0        49.5
#> 8         53.0        58.0        54.5        48.5        57.5        49.0
#> 9         46.0        51.0        49.0        48.5        43.0        49.0
#>    bin_4_rep_1 bin_4_rep_2 bin_5_rep_1 bin_5_rep_2
#> 1         46.0        41.0        34.0        50.0
#> 10        40.5        48.5        49.5        43.0
#> 2         47.0        56.5        48.5        51.5
#> 3         44.0        60.0        55.0        44.5
#> 4         48.5        50.0        59.0        47.5
#> 5         59.5        47.0        50.0        57.5
#> 6         48.5        50.5        44.0        48.0
#> 7         54.0        47.0        55.0        48.5
#> 8         49.0        49.0        51.5        53.0
#> 9         54.5        49.5        49.5        54.0